requirements.txt
torch>=2.0.1
diffusers==0.34.0
transformers>=4.37.2
accelerate
safetensors
Pillow
requests
tqdm
runpod
opencv-python
